feat: lotto list DOM 생성 로직 파일 분리
